requirements.txt
arelle-release==2.37.17
blinker==1.9.0
bottle==0.13.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
Flask-Mail==0.10.0
git-filter-repo==2.47.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
six==1.17.0
stripe==12.2.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
